+++ src/rtc/Peer.ts
@@ -119,16 +119,10 @@ export default class Peer {
   }
 
   private _config(): RTCConfiguration {
+    const { iceServers = [] } = this.options
     // bundlePolicy = "max-compat";
     // @ts-ignore
-    const config: RTCConfiguration = { sdpSemantics: 'plan-b' }
-    if (this.options.hasOwnProperty('iceServers') && this.options.iceServers) {
-      if (typeof (this.options.iceServers) === 'boolean') {
-        config.iceServers = [{ urls: ['stun:stun.l.google.com:19302'] }]
-      } else {
-        config.iceServers = this.options.iceServers
-      }
-    }
+    const config: RTCConfiguration = { sdpSemantics: 'plan-b', iceServers }
     logger.info('RTC config', config)
     return config
   }